requirements.txt
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
librosa>=0.8.1
openai>=0.27.0
nemo-toolkit[asr]>=1.8.0
scipy>=1.7.0
soundfile>=0.10.3
python-dotenv>=0.19.0
pandas>=2.0.0
scikit-learn>=1.0.0
tqdm>=4.65.0
sacrebleu>=2.5.1
rouge-score>=0.1.2
jieba>=0.42.1
opencc>=1.1.9
pangu>=4.0.6.1
pytest>=7.0.0
sounddevice>=0.4.6
